feat: Add waiting-for-ok-to-test label support
If a PR has the label `waiting-for-ok-to-test` at any point, it won't
build automatically. `needs-ok-to-test` will get automatically removed
if the author of the PR is a trusted user, so we needed a new label
here instead. The exception here is if a trusted user does `/lgtm` on
a PR with `waiting-for-ok-to-test`, in which case we'll do the same
thing as if the PR had `needs-ok-to-test` - run the contexts once, but
don't remove the label.

fixes #886

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>